requirements.txt
evtx ~= 0.8.8
python-magic ~= 0.4.27
pathvalidate ~= 3.2.1